todo de nuevo

+++ src/main/java/cl/portafolio/hostalapp/models/repository/IHabitacionRepository.java
@@ -2,7 +2,7 @@ package cl.portafolio.hostalapp.models.repository;
 
 import cl.portafolio.hostalapp.models.entity.Habitacion;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
-public interface IHabitacionRepository extends JpaRepository<Habitacion, Integer> {
+public interface IHabitacionRepository extends JpaRepository<Habitacion, Long> {
+
 }
